15.0.72

+++ src/reconstruct.c
@@ -1,6 +1,6 @@
 // ------------------------------------------------------------------
 //   reconstruct.c
-//   Copyright (C) 2019-2024 Genozip Limited. Patent Pending.
+//   Copyright (C) 2019-2025 Genozip Limited. Patent Pending.
 //   Please see terms and conditions in the file LICENSE.txt
 //
 //   WARNING: Genozip is proprietary, not open source software. Modifying the source code is strictly prohibited
@@ -565,7 +565,7 @@ void reconstruct_one_snip (VBlockP vb, ContextP snip_ctx,
     case SNIP_SPECIAL:
         ASSPIZ (snip_len >= 2, "SNIP_SPECIAL expects snip_len=%u >= 2. ctx=\"%s\"", snip_len, snip_ctx->tag_name);
                 
-        uint8_t special = snip[1] - 32; // +32 was added by SPECIAL macro
+        uint8_t special = (uint8_t)snip[1] - 32; // +32 was added by SPECIAL macro
 
         ASSPIZ (special < DTP(num_special), "Reconstructing %s requires non-existant special %s handler %u which doesn't exist", 
                 base_ctx->tag_name, dt_name (vb->data_type), special); // this happens when compiling a new special without generating dict_id_gen.h
